LDPE Sheets
LDPE is a product that has to turn out to be a must in many industries due to its wide application. These low-density polyethylenes are recognized for their high quality. these round bottomed liners act as a cover for vessels that allows you to store or deliver wet or dry chemical compounds, slurries, paste or food products effectively.